Support figuring out YTD statistics.
So far the code was figuring things out for the lifetime of the account, now
it's possible to look at statistics for the current YTD period as well.

This is useful to see YTD P/L, because TW unfortunately marks to market all
positions at year end, so YTD P/L is incorrect unless you're a professional
trader and get to mark to market your positions for tax purposes.