Allow closing more quantity than what was left open.
This is to correctly handle the case where we're long 100 shares, sell
to close 200, and find ourselves now short 100 shares. The code was
crashing because after selling 200 we had 100 remaining shares we
couldn't close. Now instead it synthesizes an opening transaction for
those 100 shares, taking the opposite of the market.